services/cartography/requirements.txt
cartography==0.87.0
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
pytest-asyncio==0.23.6
